Enable arbitrary user and host to connect to our X11 server.
The XAUTHORITY approach that we were using to enable multiple Linux
user to access Xwayland and run GUI application doesn't work by
default in docker container or when the user change the hostname
such that it differs from that of the system distro. Given all
Linux users belong to a single Windows users, we want to effectively
disable access control such that the Windows users can connect
to GUI application from within a WSL instance from any Linux users,
or container. We accomplish this by configuing Weston such that
it launches Xwayland with the "-ac" parameter to disable access
control.