Bring up of Mariner system distro.
The system distro based on Mariner is now at parity with Ubuntu. This PR does
a few things.

1)	Restore SIGCHLD signal to default handling in WSLGd as both WSLGd (for
child monitoring) and Xwayland for communication with xkbcomp (for
popen/pclose) depends on this.

2) Launch dbus-daemon with CAP_SETGID & CAP_SETUID so it can properly switch
to the messagebus UID/GID.

3) Add various components to the Mariner build image that was missing for
proper execution (xkeyboard-config, xcursor-themes, libltdl, ...) and remove
various component that were not necessary.

3) Install all binaries in /usr for both Mariner and Ubuntu distro, no more
/usr/local to avoid the possibility of accidentally getting multiple version
of binaries installed.

4) Leveraged DESTDIR to cleanly separate built binaries from the build
environment and got rid of the intermediate tar file generation and extraction.